add arduino capatability

+++ api/lexer.h
@@ -33,7 +33,7 @@ typedef struct {
     Instruction inst;
     char* text;
     char* filepath;
-    int line;
+    unsigned int line;
 } Token;
 
 typedef enum {
@@ -55,10 +55,10 @@ typedef struct {
     LexerState state;
 
     Token tokens[MAX_PROGRAM_SIZE];
-    unsigned int numTokens;
+    unsigned int  numTokens;
 
     Label labels[MAX_LABELS];
-    unsigned int numLabels;
+    unsigned short  numLabels;
 } Lexer;
 
 /// @brief An entry to represent relationship between string and enum
